refactor: simplify TreeNode handling by implementing AsRef and AsMut traits
