--- backend/server/src/domain/files/factories/mod.rs
@@ -1,1 +1,0 @@
-pub mod tree_factory;
