+++ backend/server/src/domain/files/mod.rs
@@ -1,2 +1,1 @@
 pub mod entities;
-pub mod factories;
